+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NGSI Go v0.8.2-next
 
--   Hardening: Add skipForwarding option for Orion
+-   Fix: Set Accept Header to application/json in batch request
+-   Hardening: Add skipForwarding option for Orion (#150)
 -   Fix: Fix message in cp command (#149)
 -   Update: Update test process (#148)
 -   Update: Update Orion version to 3.1.0 (#146)